Implementación CRUD completo de habitaciones con JWT y permisos

+++ src/main/java/com/hotelBackend/repository/ArticuloInventarioRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.ArticuloInventario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticuloInventarioRepository extends JpaRepository<ArticuloInventario,Integer> {
+public interface ArticuloInventarioRepository extends JpaRepository<ArticuloInventario,Long> {
 }
