+++ src/main/java/com/hotelBackend/repository/HabitacionRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.Habitacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HabitacionRepository extends JpaRepository<Habitacion, Integer> {
+public interface HabitacionRepository extends JpaRepository<Habitacion, Long> {
 }
