+++ src/main/java/com/hotelBackend/repository/MovimientoInventarioRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.MovimientoInventario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovimientoInventarioRepository extends JpaRepository<MovimientoInventario, Integer> {
+public interface MovimientoInventarioRepository extends JpaRepository<MovimientoInventario, Long> {
 }
