+++ src/main/java/com/hotelBackend/repository/PermisoRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.Permiso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PermisoRepository extends JpaRepository<Permiso, Integer> {
+public interface PermisoRepository extends JpaRepository<Permiso, Long> {
 }
