+++ src/main/java/com/hotelBackend/repository/PlanTarifarioRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.PlanTarifario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanTarifarioRepository extends JpaRepository<PlanTarifario, Integer> {
+public interface PlanTarifarioRepository extends JpaRepository<PlanTarifario, Long> {
 }
