+++ src/main/java/com/hotelBackend/repository/RegistroLimpiezaRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.RegistroLimpieza;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegistroLimpiezaRepository extends JpaRepository<RegistroLimpieza, Integer> {
+public interface RegistroLimpiezaRepository extends JpaRepository<RegistroLimpieza, Long> {
 }
