+++ src/main/java/com/hotelBackend/repository/ReservaRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.Reserva;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservaRepository extends JpaRepository<Reserva, Integer> {
+public interface ReservaRepository extends JpaRepository<Reserva, Long> {
 }
