+++ src/main/java/com/hotelBackend/repository/TicketMantenimientoRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.TicketMantenimiento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketMantenimientoRepository extends JpaRepository<TicketMantenimiento,Integer> {
+public interface TicketMantenimientoRepository extends JpaRepository<TicketMantenimiento,Long> {
 }
