+++ src/main/java/com/hotelBackend/repository/TipoHabitacionRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.TipoHabitacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoHabitacionRepository extends JpaRepository<TipoHabitacion, Integer> {
+public interface TipoHabitacionRepository extends JpaRepository<TipoHabitacion, Long> {
 }
