+++ src/main/java/com/hotelBackend/repository/TransaccionFolioRepository.java
@@ -3,5 +3,5 @@ package com.hotelBackend.repository;
 import com.hotelBackend.model.TransaccionFolio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransaccionFolioRepository extends JpaRepository<TransaccionFolio, Integer> {
+public interface TransaccionFolioRepository extends JpaRepository<TransaccionFolio, Long> {
 }
